Back/requirements.txt
cryptography==2.9.2
Flask==1.1.2
Flask-Cors==3.0.8
peewee==3.13.3
PyMySQL==0.9.3
pyjwt==1.7.1